fixedissue with timezone offset by applying timezone like zoneId = ZoneId.of(dailyWeather.timezone) also fixed crash on API<=25 with core library desugaring not enabled //TODO remove unused timezoneOffset fix 0 results when timezone on device and city timezone not equal

+++ app/src/main/java/com/example/weatherforecast/response/ForecastResponse.kt
@@ -18,7 +18,7 @@ data class ForecastResponse(
     @SerializedName("timezone")
     val timezone: String,
     @SerializedName("timezone_offset")
-    val timezoneOffset: Int
+    val timezoneOffset: Double
 )
 data class Current(
     @SerializedName("clouds")
